backend/requirements.txt
# ForgeFlow Lite - Backend Dependencies
# Phase 1: 기반 구축

# ============================================
# Web Framework
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================
# Database
# ============================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ============================================
# Caching
# ============================================
redis==5.0.1
hiredis==2.3.2  # Redis 성능 최적화

# ============================================
# Data Validation
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# Environment & Configuration
# ============================================
python-dotenv==1.0.0

# ============================================
# HTTP Client & File Handling
# ============================================
httpx==0.25.1
aiofiles==23.2.1

# ============================================
# Data Processing (Phase 4용, 사전 설치)
# ============================================
pandas==2.1.3

# ============================================
# AI/ML (Phase 2용, 사전 설치)
# ============================================
google-generativeai==0.8.3

# ============================================
# Testing (Optional)
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
